// airways/src/app/auth/models/login-form-data-type.model.ts
export type LoginFormDataType = {
  email: string;
  password: string;
};

export type RegisterFormDataType = {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  bthDate: Date;
  gender: string;
  country: string;
  phoneNumber: string;
  citizenship: string;
  acceptTerms: boolean;
};